requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
tensorrt>=8.5.0
onnx>=1.12.0
ultralytics>=8.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
pyyaml>=6.0
jupyter>=1.0.0
cmake>=3.18.0
